+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   in CamelCase to snake\_case, kebab-case, or with a custom separator.
 * Remove redundant metadata about the query nature when logging the query
 * Add `selectOneWhereIn` that can match a row's column in a value of user-provided values
+* Add `selectOrderBy` to specify a vector of sorting specs (field + sorting keyword) with your select.
+* Fix primary key detection in generic deriving
 
 ## 0.0.1.0 -- 2021-11-05